requirements.txt
asgiref==3.11.1
authease==3.2.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.11
inflection==0.5.1
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
